Added code for Stock Class
Currently the code does not build (need EMA.h, MACD.h
etc. and also check for any bugs) but we can work on
this tomorrow when we have all the code

+++ PairsTrading.cpp
@@ -13,7 +13,7 @@ using namespace std;
 
 // dummy risk management and order placement modules (not part of this class?)
 
-float getInvestmentAmount (string order, float diff, int count) {
+double getInvestmentAmount (string order, double diff, int count) {
 
     /*
         Consider:
@@ -30,13 +30,14 @@ float getInvestmentAmount (string order, float diff, int count) {
 }
 
 // requires order management module
-void placeOrder(string order, string tick, float amount) {
+void placeOrder(string order, string tick, double amount) {
 }
 
 // PairsTrading Class
 
 PairsTrading::PairsTrading(Stock * _s1, Stock * _s2) {
-    s1 = _s1;    
+	// should be "new"ed before calling this constructor
+	s1 = _s1;    
     s2 = _s2;
 }
 
@@ -118,7 +119,7 @@ void PairsTrading::State3() {
     if (s1Data.currentEMA < s1Data.EMAatDivergence) {
         // part of Risk management module (separate class? inputs?)
         // or we could make one risk management module for just this algorithm
-        float buyAmount = getInvestmentAmount ("buy",
+        double buyAmount = getInvestmentAmount ("buy",
                               s1Data.currentEMA - s1Data.EMAatDivergence,
                               divergedCount
                               );
@@ -127,7 +128,7 @@ void PairsTrading::State3() {
         placeOrder("buy", s1->getTick(), buyAmount);
     }
     else { // stock 1 is rising
-        float sellAmount = getInvestmentAmount ("sell",
+        double sellAmount = getInvestmentAmount ("sell",
                               s1Data.currentEMA - s1Data.EMAatDivergence,
                               divergedCount
                               );
@@ -138,14 +139,14 @@ void PairsTrading::State3() {
     if (s2Data.currentEMA < s2Data.EMAatDivergence) {
         // part of Risk management module (separate class? inputs?)
         // or we could make one risk management module for just this algorithm
-        float buyAmount = getInvestmentAmount ("buy",
+        double buyAmount = getInvestmentAmount ("buy",
                               s2Data.currentEMA - s2Data.EMAatDivergence,
                               divergedCount
                               );
         placeOrder("buy", s2->getTick(), buyAmount);
     }
     else { // stock 2 is rising
-        float sellAmount = getInvestmentAmount ("sell",
+        double sellAmount = getInvestmentAmount ("sell",
                               s2Data.currentEMA - s2Data.EMAatDivergence,
                               divergedCount
                               );
@@ -192,7 +193,7 @@ void PairsTrading::start() {
 }
 
 // currently unused (for testing only)
-float PairsTrading::getDecision() {
+double PairsTrading::getDecision() {
     if ( s1->getMACD() > s1->getEMA() ) {
          cout << "Stock 1: useMACD" << endl;
     }
